autoscroll added, character mechanics, platforms
I am really stuck at this one, check if you have an idea how to fix it otherwise I am gonna ask teacher. The camera is moving but the scene doesn't which means we can't check if the platforms are out of the scene. If you have another idea how to do this it would be great.